Move the plugin files to plugins directory
