+++ SimG4Core/PrintGeomInfo/plugins/PrintMaterialBudgetInfo.cc
@@ -1,10 +1,11 @@
-#include "SimG4Core/PrintGeomInfo/interface/PrintMaterialBudgetInfo.h"
-
 #include "SimG4Core/Notification/interface/BeginOfJob.h"
 #include "SimG4Core/Notification/interface/BeginOfRun.h"
+#include "SimG4Core/Notification/interface/Observer.h"
+#include "SimG4Core/Watcher/interface/SimWatcher.h"
 
 #include "FWCore/Framework/interface/EventSetup.h"
 #include "FWCore/Framework/interface/ESHandle.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "Geometry/Records/interface/IdealGeometryRecord.h"
 #include "DetectorDescription/Core/interface/DDCompactView.h"
 #include "DetectorDescription/Core/interface/DDFilter.h"
@@ -20,6 +21,7 @@
 #include "G4LogicalVolume.hh"
 #include "G4VSolid.hh"
 #include "G4Material.hh"
+#include "G4NavigationHistory.hh"
 #include "G4Track.hh"
 #include "G4VisAttributes.hh"
 #include "G4UserLimits.hh"
@@ -27,7 +29,58 @@
 #include "G4UnitsTable.hh"
 #include "Randomize.hh"
 
+#include <iostream>
+#include <fstream>
 #include <set>
+#include <vector>
+#include <string>
+
+typedef std::map<G4VPhysicalVolume*, G4VPhysicalVolume*, std::less<G4VPhysicalVolume*> > mpvpv;
+typedef std::multimap<G4LogicalVolume*, G4VPhysicalVolume*, std::less<G4LogicalVolume*> > mmlvpv;
+
+class PrintMaterialBudgetInfo : public SimWatcher,
+                                public Observer<const BeginOfJob*>,
+                                public Observer<const BeginOfRun*> {
+public:
+  PrintMaterialBudgetInfo(edm::ParameterSet const& p);
+  ~PrintMaterialBudgetInfo() override;
+
+private:
+  void update(const BeginOfJob* job) override{};
+  void update(const BeginOfRun* run) override;
+  void dumpHeader(std::ostream& out = std::cout);
+  void dumpLaTeXHeader(std::ostream& out = std::cout);
+  void dumpHierarchyLeaf(G4VPhysicalVolume* pv,
+                         G4LogicalVolume* lv,
+                         unsigned int leafDepth,
+                         std::ostream& weightOut = std::cout,
+                         std::ostream& texOut = std::cout);
+  void printInfo(G4VPhysicalVolume* pv,
+                 G4LogicalVolume* lv,
+                 unsigned int leafDepth,
+                 std::ostream& weightOut = std::cout,
+                 std::ostream& texOut = std::cout);
+  void dumpElementMassFraction(std::ostream& elementOut = std::cout);
+  void dumpLaTeXFooter(std::ostream& out = std::cout);
+
+private:
+  std::string name;
+  int nchar;
+  mpvpv thePVTree;
+  G4VPhysicalVolume* theTopPV;
+  G4NavigationHistory fHistory;
+  bool volumeFound;
+  unsigned int levelFound;
+  std::ofstream weightOutputFile;
+  std::ofstream elementOutputFile;
+  std::ofstream texOutputFile;
+  std::vector<std::string> elementNames;
+  std::vector<double> elementTotalWeight;
+  std::vector<double> elementWeightFraction;
+  //
+  std::string stringLaTeXUnderscore(std::string stringname);
+  std::string stringLaTeXSuperscript(std::string stringname);
+};
 
 PrintMaterialBudgetInfo::PrintMaterialBudgetInfo(const edm::ParameterSet& p) {
   name = p.getUntrackedParameter<std::string>("Name", "*");
@@ -303,3 +356,8 @@ std::string PrintMaterialBudgetInfo::stringLaTeXSuperscript(std::string stringna
 
   return stringoutput;
 }
+
+#include "SimG4Core/Watcher/interface/SimWatcherFactory.h"
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+
+DEFINE_SIMWATCHER(PrintMaterialBudgetInfo);
